move poll logic into new file
